winpthreads: Silence warning about dropping a const qualifier
clang-cl warns about dropping the const-qualifier of the value given
to pthread_setspecific. It's however valid to do so, as the optional
destructor specified with pthread_key_create takes a void* value (not
const). If the value has been dynamically allocated, then casting away
the const is valid. If the value was truly const, then the user likely
hasn't defined any destructor for it, and casting the const away,
although technically UB, should be harmless.

Using memcpy prevents the warning because it 'forgets' an
indirection. It's equivalent to:

t->keyval[key] = *(void * const *)(const void *)&value;

Signed-off-by: Antonin Décimo <antonin@tarides.com>