winpthreads: Avoid triggering the bad-function-cast warning
The bad-function-cast warning triggers when a function call returning
an integer type (resp. a pointer) is cast to a pointer (resp. an
integer type), even when the integer is wide enough to hold a
pointer (the case here for intptr_t and uintptr_t).

The warning can be avoided by adding an explicit cast of the actual
function return type.

Signed-off-by: Antonin Décimo <antonin@tarides.com>